refactor: implementation of create_mac_packet is moved from the header file to the .c file
